Compute empirical mortality rate

Calculate the empirical mortality rate over a given interval.

Usage

mortality_rate(lifelihoodData, interval_width, max_time = NULL, bygroup = TRUE)

Arguments

lifelihoodData

lifelihoodData object created with lifelihoodData().

interval_width

The interval width used to calculate the mortality rate. For instance, if the time unit for deaths in the original dataset is days and interval_width is set to 10, the mortality rate will be calculated every 10 days for each group.

max_time

The maximum time for calculating the mortality rate. If set to NULL, the time of the last observed death is used.

bygroup

Whether to calculate mortality rates by subgroup (default is TRUE). If FALSE, calculates a single overall mortality rate.

Value

A dataframe with 3 columns: Interval (time interval, based on interval_width value), Group (identifier of a given subgroup, or "Overall" if bygroup = FALSE), and MortalityRate (mortality rate at this time).

Examples

df <- read.csv(here::here("data/fake_sample.csv"))
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

mort_df <- mortality_rate(dataLFH, interval_width = 2)
head(mort_df)
#>   Interval Group MortalityRate
#> 1        2     1             0
#> 2        4     1             0
#> 3        6     1             1
#> 4        8     1             1
#> 5       10     1             1
#> 6       12     1             1

mort_df <- mortality_rate(
  dataLFH,
  interval_width = 2,
  bygroup = FALSE,
  max_time = 170
)
head(mort_df)
#>   Interval MortalityRate
#> 1        2     0.2307692
#> 2        4     0.0000000
#> 3        6     0.4000000
#> 4        8     0.0000000
#> 5       10     0.1666667
#> 6       12     0.2000000