pkgdown/assets/header.html

Prediction from a lifelihood model at new data values

S3 method used to make predictions using lifelihoodResults objects, i.e. results from lifelihood().

Usage

prediction(
  object,
  parameter_name,
  newdata = NULL,
  mcmc.fit = FALSE,
  keep_mcmc_samples = FALSE,
  type = c("link", "response"),
  se.fit = FALSE,
  .warning_ratio_male = TRUE
)

Arguments

object

output of lifelihood()

parameter_name

A string, or vector of strings, specifying the name(s) of the parameter(s) for which to make the prediction. Each name must be one of unique(lifelihoodResults$effects$parameter).

newdata

Data for prediction. If absent, predictions are for each individual in the original dataset provided by the user.

mcmc.fit

Whether to return predictions based on the fitted MCMC samples. Requires fitting the model with MCMC > 0.

keep_mcmc_samples

Whether or not to also retrieve MCMC samples in output. If TRUE, output is a list with 2 elements: pred and mcmc_samples.

type

The type of the predicted value: if type="response," predictions are on the original data scale; if type="link," predictions are on the lifelihood scale.

se.fit

Whether or not to include standard errors in the prediction (computed on the response scale using the delta method).

.warning_ratio_male

Whether to warn when male predictions for expt_death are identical to female predictions on the link scale.

Value

For a single parameter, a vector or list containing the predicted values for the parameter. For multiple parameters, a data frame with one column per parameter, or a list containing parameter-prefixed data frames when MCMC samples are kept.

Examples

df <- datapierrick |>
  dplyr::mutate(
    par = as.factor(par),
    spore = as.factor(spore)
  )

clutchs <- generate_clutch_vector(28)

dataLFH <- as_lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("par", "spore"),
  matclutch = FALSE,
  dist = c(mortality = "wei", maturity = "gam", reproduction = "lgn")
)

results <- lifelihood(
  lifelihoodData = dataLFH,
  config = list(
    mortality = list(expt_death = "par + spore", survival_param2 = 1)
  ),
  seeds = c(1, 2, 3, 4),
  raise_estimation_warning = FALSE
)

prediction(results, "expt_death")
#> Lifelihood parameter estimate(s) for males are identical to that of females. Use type='response', to get the right parameter estimate(s) for males on the response scale.
#>   [1] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#>   [8] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#>  [15] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#>  [22] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#>  [29] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -1.702634 -2.029652
#>  [36] -1.702634 -1.702634 -2.029652 -1.702634 -1.702634 -1.702634 -1.702634
#>  [43] -1.702634 -1.702634 -2.029652 -1.702634 -2.029652 -2.029652 -2.029652
#>  [50] -2.029652 -2.029652 -1.702634 -2.029652 -2.029652 -1.702634 -1.702634
#>  [57] -2.029652 -1.702634 -1.702634 -1.702634 -1.702634 -1.702634 -2.029652
#>  [64] -1.702634 -1.702634 -2.029652 -2.029652 -2.029652 -1.735566 -2.029652
#>  [71] -1.735566 -2.029652 -1.735566 -2.029652 -1.735566 -1.735566 -1.735566
#>  [78] -1.735566 -1.735566 -1.735566 -2.029652 -2.029652 -1.735566 -2.029652
#>  [85] -1.735566 -1.735566 -1.735566 -2.029652 -1.735566 -1.735566 -2.029652
#>  [92] -1.735566 -1.735566 -1.735566 -1.735566 -1.735566 -1.735566 -1.735566
#>  [99] -1.735566 -2.029652 -1.735566 -1.735566 -1.735566 -1.657905 -1.657905
#> [106] -1.657905 -2.029652 -2.029652 -2.029652 -2.029652 -1.657905 -1.657905
#> [113] -1.657905 -1.657905 -2.029652 -1.657905 -2.029652 -1.657905 -2.029652
#> [120] -1.657905 -2.029652 -2.029652 -2.029652 -1.657905 -2.029652 -1.657905
#> [127] -1.657905 -1.657905 -1.657905 -1.657905 -1.657905 -1.657905 -1.657905
#> [134] -1.657905 -1.657905 -1.657905 -1.657905 -1.657905 -2.029652 -2.029652
#> [141] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [148] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [155] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [162] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [169] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [176] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [183] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [190] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [197] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [204] -2.029652 -2.029652 -2.029652 -2.029652 -1.735566 -1.735566 -1.735566
#> [211] -1.735566 -1.735566 -1.735566 -1.735566 -1.735566 -1.735566 -1.735566
#> [218] -1.735566 -1.735566 -2.029652 -1.735566 -1.735566 -2.029652 -2.029652
#> [225] -1.735566 -1.735566 -2.029652 -2.029652 -2.029652 -1.735566 -1.735566
#> [232] -2.029652 -1.735566 -2.029652 -1.735566 -2.029652 -2.029652 -1.735566
#> [239] -2.029652 -2.029652 -2.029652 -1.657905 -1.657905 -1.657905 -1.657905
#> [246] -1.657905 -1.657905 -1.657905 -1.657905 -1.657905 -1.657905 -1.657905
#> [253] -1.657905 -1.657905 -1.657905 -1.657905 -2.029652 -1.657905 -1.657905
#> [260] -1.657905 -1.657905 -2.029652 -2.029652 -1.657905 -1.657905 -2.029652
#> [267] -1.657905 -1.657905 -1.657905 -2.029652 -2.029652 -1.657905 -2.029652
#> [274] -1.657905 -1.657905 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [281] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [288] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [295] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [302] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [309] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [316] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [323] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [330] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [337] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [344] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [351] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [358] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [365] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [372] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [379] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [386] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [393] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [400] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [407] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [414] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [421] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [428] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [435] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [442] -2.029652 -2.029652 -2.029652 -2.029652 -1.702634 -1.702634 -1.702634
#> [449] -1.702634 -1.702634 -1.702634 -2.029652 -1.702634 -1.702634 -1.702634
#> [456] -2.029652 -1.702634 -1.702634 -1.702634 -1.702634 -1.702634 -2.029652
#> [463] -2.029652 -1.702634 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [470] -1.702634 -1.702634 -1.702634 -1.702634 -1.702634 -1.702634 -2.029652
#> [477] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [484] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [491] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [498] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [505] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [512] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [519] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [526] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [533] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [540] -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652 -2.029652
#> [547] -2.029652 -2.029652 -2.029652 -2.029652
prediction(results, "expt_death", type = "response")
#>   [1] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#>   [9] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#>  [17] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#>  [25] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#>  [33] 37.62528 49.93623 37.62528 49.93623 49.93623 37.62528 49.93623 49.93623
#>  [41] 49.93623 49.93623 49.93623 49.93623 37.62528 49.93623 37.62528 37.62528
#>  [49] 37.62528 37.62528 37.62528 49.93623 37.62528 37.62528 49.93623 49.93623
#>  [57] 37.62528 49.93623 49.93623 49.93623 49.93623 49.93623 37.62528 49.93623
#>  [65] 49.93623 37.62528 37.62528 37.62528 48.56102 37.62528 48.56102 37.62528
#>  [73] 48.56102 37.62528 48.56102 48.56102 48.56102 48.56102 48.56102 48.56102
#>  [81] 37.62528 37.62528 48.56102 37.62528 48.56102 48.56102 48.56102 37.62528
#>  [89] 48.56102 48.56102 37.62528 48.56102 48.56102 48.56102 48.56102 48.56102
#>  [97] 48.56102 48.56102 48.56102 37.62528 48.56102 48.56102 48.56102 51.85490
#> [105] 51.85490 51.85490 37.62528 37.62528 37.62528 37.62528 51.85490 51.85490
#> [113] 51.85490 51.85490 37.62528 51.85490 37.62528 51.85490 37.62528 51.85490
#> [121] 37.62528 37.62528 37.62528 51.85490 37.62528 51.85490 51.85490 51.85490
#> [129] 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490
#> [137] 51.85490 51.85490 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [145] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [153] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [161] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [169] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [177] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [185] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [193] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [201] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 48.56102
#> [209] 48.56102 48.56102 48.56102 48.56102 48.56102 48.56102 48.56102 48.56102
#> [217] 48.56102 48.56102 48.56102 37.62528 48.56102 48.56102 37.62528 37.62528
#> [225] 48.56102 48.56102 37.62528 37.62528 37.62528 48.56102 48.56102 37.62528
#> [233] 48.56102 37.62528 48.56102 37.62528 37.62528 48.56102 37.62528 37.62528
#> [241] 37.62528 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490
#> [249] 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490 51.85490
#> [257] 37.62528 51.85490 51.85490 51.85490 51.85490 37.62528 37.62528 51.85490
#> [265] 51.85490 37.62528 51.85490 51.85490 51.85490 37.62528 37.62528 51.85490
#> [273] 37.62528 51.85490 51.85490 37.62528 37.62528 37.62528 37.62528 37.62528
#> [281] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [289] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [297] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [305] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [313] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [321] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [329] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [337] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [345] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [353] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [361] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [369] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [377] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [385] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [393] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [401] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [409] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [417] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [425] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [433] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [441] 37.62528 37.62528 37.62528 37.62528 37.62528 49.93623 49.93623 49.93623
#> [449] 49.93623 49.93623 49.93623 37.62528 49.93623 49.93623 49.93623 37.62528
#> [457] 49.93623 49.93623 49.93623 49.93623 49.93623 37.62528 37.62528 49.93623
#> [465] 37.62528 37.62528 37.62528 37.62528 37.62528 49.93623 49.93623 49.93623
#> [473] 49.93623 49.93623 49.93623 37.62528 37.62528 37.62528 37.62528 37.62528
#> [481] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [489] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [497] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [505] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [513] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [521] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [529] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [537] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528
#> [545] 37.62528 37.62528 37.62528 37.62528 37.62528 37.62528

# predict on new data
newdata <- data.frame(
  par = c(0, 1, 2, 0, 1, 2),
  spore = c(0, 1, 0, 1, 0, 1)
)
newdata$par <- factor(newdata$par)
newdata$spore <- factor(newdata$spore)
prediction(results, "expt_death", newdata)
#> [1] -2.029652 -1.735566 -1.865740 -1.778890 -1.986328 -1.614977
prediction(results, "expt_death", newdata, type = "response")
#> [1] 37.62528 48.56102 43.42836 46.79951 39.09016 53.75201