Run lifelihood program

Run lifelihood program in console mode

Usage

run_lifelihood_console(
  infile = "100%mort_Pierrick211genoparinteraction.txt",
  customfile = "custom.txt",
  GbyG = 0,
  MCMC = 0,
  interval = 25,
  SEcal = 0,
  saveprobevent = 0,
  fitness = 0,
  r = 0,
  seed1 = 12,
  seed2 = 13,
  seed3 = 14,
  seed4 = 15,
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001
)

Arguments

infile

file with the model and data to be fitted

customfile

file with the min and max boudaries for each parameter

GbyG

Option to fit the full factorail model with all the interactions between each of the factors

MCMC

Perform MCMC sampling of the parameter after convergence to estimate their 95% confidence interval

interval

TBD - Check the actual meaning

SEcal

Compute the standard error of eahc parameter using the Hessian matrix

saveprobevent

TBD - Check the actual meaning

fitness

Reparametrize the model with one parameter as the lifetime reproductive success

r

Reparametrize the model with one parameter as the intrinsic rate of increase

seed1

first seed used so that random values can found again (https://en.wikipedia.org/wiki/Multiply-with-carry_pseudorandom_number_generator)

seed2

second seed used so that random values can found again

seed3

third seed used so that random values can found again

seed4

fourth seed used so that random values can found again

ntr

number of thread for the paralelisation ?

nst

TBD - Check the actual meaning

To

Initial temperature for the simulated annealing

Tf

Initial temperature for the simulated annealing

climbrate

Rate for the simulated annealing ?

precision

TBD - Check the actual meaning

Value

dataset with the simulated hatch rate

Examples

run_lifelihood_console(infile="100%mort_Pierrick211genoparinteraction.txt", ntr=10, To=50, Tf=1, seed2=34)
#> [1] "\\Users\\josephbarbier\\Desktop\\Lifelihood\\docs\\reference\\100%mort_Pierrick211genoparinteraction.txt \\Users\\josephbarbier\\Desktop\\Lifelihood\\docs\\reference\\custom.txt 0 0 25 0 0 0 0 12 34 14 15 10 2 50 1 1 0.001"
#> arguments 'show.output.on.console', 'minimized' and 'invisible' are for Windows only
#> Warning: error in running command