Summary function to be used with the output of lifelihood()

S3 method to display main results of the lifelihood program.

Usage

# S3 method for class 'lifelihoodResults'
summary(object, ...)

Arguments

object

output of lifelihood()

...

Ignored

Examples

library(lifelihood)
library(tidyverse)

df <- fakesample |>
  mutate(
    geno = as.factor(geno),
    type = as.factor(type)
  )
head(df)
#>   type geno sex_start sex_end sex mat_start mat_end clutch_start1 clutch_end1
#> 1    1    0         0    1000   0         0    1000            NA          NA
#> 2    2    0         0    1000   0         2       3             4           3
#> 3    0    1         0    1000   0         3       4             2           4
#> 4    0    1         0    1000   0         3       4             2           4
#> 5    0    1         0    1000   0         3       4             2           4
#> 6    1    0         0    1000   0         0    1000            NA          NA
#>   clutch_size1 clutch_start2 clutch_end2 clutch_size2 death_start death_end
#> 1           NA            NA          NA           NA         0.1         1
#> 2            4             2           4            5         9.0        10
#> 3            5            NA          NA           NA         5.0         6
#> 4            5            NA          NA           NA         5.0         6
#> 5            5            NA          NA           NA         5.0         6
#> 6           NA            NA          NA           NA         0.1         1

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

results <- lifelihood(
  lifelihoodData = dataLFH,
  path_config = get_config_path("config"),
  seeds = c(1, 2, 3, 4),
  raise_estimation_warning = FALSE
)
#> lifelihood Pascal program status code 0
summary(results)
#> LIFELIHOOD RESULTS
#> 
#> Likelihood:
#> [1] -594.6025
#> 
#> Effects:
#>                                       name  estimation stderror
#> 1                           int_expt_death -11.9449429        0
#> 2                    eff_expt_death_geno_1 -12.2167414        0
#> 3                    eff_expt_death_type_1 -17.4578138        0
#> 4                    eff_expt_death_type_2  -9.5788906        0
#> 5                       int_survival_shape -12.4849939        0
#> 6                eff_survival_shape_geno_1 -19.7491969        0
#> 7                eff_survival_shape_type_1 -17.3748815        0
#> 8                eff_survival_shape_type_2   2.6682119        0
#> 9         eff_survival_shape_type_1:geno_1  -6.5460851        0
#> 10        eff_survival_shape_type_2:geno_1 -14.5649205        0
#> 11                    int_ratio_expt_death  -7.1006194        0
#> 12             eff_ratio_expt_death_geno_1  10.4608933        0
#> 13                          int_prob_death -10.4279891        0
#> 14                   eff_prob_death_geno_1 -11.3766401        0
#> 15                       int_expt_maturity   0.4246114        0
#> 16                eff_expt_maturity_geno_1  -0.5987078        0
#> 17                eff_expt_maturity_type_1  10.2893645        0
#> 18                eff_expt_maturity_type_2  -1.0230025        0
#> 19                      int_maturity_shape -19.5699810        0
#> 20               eff_maturity_shape_geno_1  -7.1923330        0
#> 21                 int_ratio_expt_maturity  14.5293003        0
#> 22          eff_ratio_expt_maturity_geno_1   8.0512975        0
#> 23          eff_ratio_expt_maturity_type_1  -2.0374967        0
#> 24          eff_ratio_expt_maturity_type_2  -7.0457527        0
#> 25                   int_expt_reproduction  -1.5884860        0
#> 26            eff_expt_reproduction_geno_1  11.6826788        0
#> 27                               int_pontn  -8.8326239        0
#> 28                        eff_pontn_geno_1 -19.9999601        0
#> 29               int_increase_death_hazard   9.3567346        0
#> 30        eff_increase_death_hazard_geno_1 -17.1122105        0
#> 31        eff_increase_death_hazard_type_1   4.6742563        0
#> 32        eff_increase_death_hazard_type_2  10.6544722        0
#> 33 eff_increase_death_hazard_type_1:geno_1  17.7414034        0
#> 34 eff_increase_death_hazard_type_2:geno_1   3.1766257        0
#> 35                  int_tof_reduction_date  -5.0608854        0
#> 36           eff_tof_reduction_date_geno_1  17.0587281        0
#> 37            int_increase_tof_n_offspring  -7.7631429        0
#> 38     eff_increase_tof_n_offspring_geno_1   1.0542894        0
#> 39                 int_lin_decrease_hazard   3.7536999        0
#> 40          eff_lin_decrease_hazard_geno_1   6.3045810        0
#> 41          eff_lin_decrease_hazard_type_1  -9.9093761        0
#> 42          eff_lin_decrease_hazard_type_2  10.2707608        0
#> 43   eff_lin_decrease_hazard_type_1:geno_1  10.5169971        0
#> 44   eff_lin_decrease_hazard_type_2:geno_1 -16.5269890        0
#> 45                     int_quad_senescence  16.0191401        0
#> 46              eff_quad_senescence_geno_1 -17.9411463        0
#> 47                int_quad_decrease_hazard -13.4062494        0
#> 48         eff_quad_decrease_hazard_geno_1  16.4551575        0
#> 49             int_quad_change_n_offspring   2.8626095        0
#> 50      eff_quad_change_n_offspring_geno_1  -1.0616118        0
#> 51                     int_tof_n_offspring  -5.2294123        0
#>                   parameter        kind        event
#> 1                expt_death   intercept    mortality
#> 2                expt_death coefficient    mortality
#> 3                expt_death coefficient    mortality
#> 4                expt_death coefficient    mortality
#> 5            survival_shape   intercept    mortality
#> 6            survival_shape coefficient    mortality
#> 7            survival_shape coefficient    mortality
#> 8            survival_shape coefficient    mortality
#> 9            survival_shape coefficient    mortality
#> 10           survival_shape coefficient    mortality
#> 11         ratio_expt_death   intercept    mortality
#> 12         ratio_expt_death coefficient    mortality
#> 13               prob_death   intercept    mortality
#> 14               prob_death coefficient    mortality
#> 15            expt_maturity   intercept     maturity
#> 16            expt_maturity coefficient     maturity
#> 17            expt_maturity coefficient     maturity
#> 18            expt_maturity coefficient     maturity
#> 19           maturity_shape   intercept     maturity
#> 20           maturity_shape coefficient     maturity
#> 21      ratio_expt_maturity   intercept     maturity
#> 22      ratio_expt_maturity coefficient     maturity
#> 23      ratio_expt_maturity coefficient     maturity
#> 24      ratio_expt_maturity coefficient     maturity
#> 25        expt_reproduction   intercept reproduction
#> 26        expt_reproduction coefficient reproduction
#> 27                    pontn   intercept reproduction
#> 28                    pontn coefficient reproduction
#> 29    increase_death_hazard   intercept reproduction
#> 30    increase_death_hazard coefficient reproduction
#> 31    increase_death_hazard coefficient reproduction
#> 32    increase_death_hazard coefficient reproduction
#> 33    increase_death_hazard coefficient reproduction
#> 34    increase_death_hazard coefficient reproduction
#> 35       tof_reduction_date   intercept reproduction
#> 36       tof_reduction_date coefficient reproduction
#> 37 increase_tof_n_offspring   intercept reproduction
#> 38 increase_tof_n_offspring coefficient reproduction
#> 39      lin_decrease_hazard   intercept reproduction
#> 40      lin_decrease_hazard coefficient reproduction
#> 41      lin_decrease_hazard coefficient reproduction
#> 42      lin_decrease_hazard coefficient reproduction
#> 43      lin_decrease_hazard coefficient reproduction
#> 44      lin_decrease_hazard coefficient reproduction
#> 45          quad_senescence   intercept reproduction
#> 46          quad_senescence coefficient reproduction
#> 47     quad_decrease_hazard   intercept reproduction
#> 48     quad_decrease_hazard coefficient reproduction
#> 49  quad_change_n_offspring   intercept reproduction
#> 50  quad_change_n_offspring coefficient reproduction
#> 51          tof_n_offspring   intercept reproduction
#>