Execution of the compiled files

Run lifelihood program in console mode.

Usage

execute_bin(
  path_input_data,
  path_param_bounds,
  group_by_group,
  MCMC,
  interval,
  SEcal,
  saveprobevent,
  fitness,
  r,
  seed1,
  seed2,
  seed3,
  seed4,
  ntr,
  nst,
  To,
  Tf,
  climbrate,
  precision
)

Arguments

path_input_data

Path to the input text file with the model and data to be fitted.

path_param_bounds

Path to the parameter boundaries text file with the min and max boudaries for each parameter.

group_by_group

Boolean option to fit the full factorail model with all the interactions between each of the factors

MCMC

Perform MCMC sampling of the parameter after convergence to estimate their 95% confidence interval

interval

TBD - Check the actual meaning

SEcal

If TRUE, Lifelihood computes the standard error of each parameter using the Hessian matrix (output with value of -1 if standard error cannot be computed due to singularity of the Hessian matrix)

saveprobevent

TBD - Check the actual meaning

fitness

Reparametrize the model with one parameter as the lifetime reproductive success

r

Reparametrize the model with one parameter as the intrinsic rate of increase

seed1

First seed number used to reproduce results (same seed = same results).

seed2

Second seed number used to reproduce results (same seed = same results).

seed3

Third seed number used to reproduce results (same seed = same results).

seed4

4th seed number used to reproduce results (same seed = same results).

ntr

Number of thread for the paralelisation ?

nst

TBD - Check the actual meaning

To

Initial temperature for the simulated annealing

Tf

Initial temperature for the simulated annealing

climbrate

Rate for the simulated annealing ?

precision

TBD - Check the actual meaning

Value

NULL. This function writes an output file, that will then be parsed by parse_output().