Individual life history modelling

Main function of the lifelihood program. Provides the joined likelihood of all the events in an individual life-history (time of maturity, reproductive events, death).

Usage

lifelihood(
  df,
  path_config,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  clutchs,
  death_start,
  death_end,
  models,
  covariates,
  matclutch = FALSE,
  matclutch_size = NULL,
  param_range_df = NULL,
  group_by_group = FALSE,
  MCMC = 0,
  interval = 25,
  SEcal = 0,
  saveprobevent = 0,
  fitness = 0,
  r = 0,
  seeds = c(12, 13, 14, 15),
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001,
  right_censoring_date = 1000,
  critical_age = 20,
  ratiomax = 10
)

Arguments

df

Dataframe with the data of life history. It should have one row per life history / observation.

path_config

Path to the configuration file (YAML format).

sex

Column name containing the sex of the observations.

sex_start

Column name containing the first date of the interval in which the sex was determined.

sex_end

Column name containing the second date of the interval in which the sex was determined.

maturity_start

Column name containing the first date of the interval in which the maturity was determined.

maturity_end

Column name containing the second date of the interval in which the maturity was determined.

clutchs

Vector containing the names of the clutch columns. The order should be: first clutch first date, first clutch second date, first clutch clutch size, second clutch first date, first clutch second date, second clutch clutch size, and so on. If the observation with the most clutches is, for example, 10, then the vector must be of size 10 x 3 = 30 (3 elements per clutch: first date, second date and size).

death_start

Column name containing the first date of the interval in which the death was determined.

death_end

Column name containing the second date of the interval in which the death was determined.

models

Vector of characters with the name of the statistical law to use. Must be of length 3 and each element must be in "wei", "exp", "gam" or "lgn". The first one is used for maturity, the second one is used for clutchs and the third one for death.

covariates

Vector containing the names of the covariates.

matclutch

Whether the maturity event (designated by maturity_start and maturity_end) is a clutch event or not. If TRUE, must specify the matclutch_size argument.

matclutch_size

Column name containing the size of the clutch for the maturity event. Only used (and required) if matclutch is TRUE.

param_range_df

Dataframe with the parameter ranges/boundaries/boundaries

group_by_group

Option to fit the full factorail model with all the interactions between each of the factors

MCMC

Perform MCMC sampling of the parameter after convergence to estimate their 95% confidence interval

interval

TBD - Check the actual meaning

SEcal

Compute the standard error of each parameter using the Hessian matrix

saveprobevent

TBD - Check the actual meaning

fitness

Reparametrize the model with one parameter as the lifetime reproductive success

r

Reparametrize the model with one parameter as the intrinsic rate of increase

seeds

Numbers used to reproduce results (same seeds = same results)

ntr

Number of thread for the paralelisation ?

nst

TBD - Check the actual meaning

To

Initial temperature for the simulated annealing

Tf

Initial temperature for the simulated annealing

climbrate

Rate for the simulated annealing ?

precision

TBD - Check the actual meaning

right_censoring_date

Time (integer) point at which a subject’s data is censored. This means that for subjects who do not experience the event of interest (e.g., death, failure) by this date, their data is considered censored. In practice, choose a value much larger than the maximum longevity seen in the data. (CURRENTLY IGNORED)

critical_age

Critical age (integer) below which life histories are not followed individually. (CURRENTLY IGNORED)

ratiomax

Maximum ratio (integer) between number of offspring of last and first reproduction events. Cannot be greater than ratiomax. (CURRENTLY IGNORED)